+++ plugin/src/main/kotlin/spp/jetbrains/sourcemarker/command/status/ui/BreakpointStatusBar.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package spp.jetbrains.sourcemarker.command.ui.status;
+package spp.jetbrains.sourcemarker.command.status.ui;
 
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.editor.Editor;
@@ -23,6 +23,7 @@ import com.intellij.openapi.editor.event.VisibleAreaListener;
 import com.intellij.openapi.editor.impl.EditorImpl;
 import com.intellij.ui.DocumentAdapter;
 import com.intellij.ui.JBColor;
+import com.intellij.ui.JBIntSpinner;
 import com.intellij.ui.components.JBScrollPane;
 import com.intellij.ui.table.JBTable;
 import com.intellij.util.ui.ColumnInfo;
@@ -30,6 +31,7 @@ import com.intellij.util.ui.JBUI;
 import com.intellij.util.ui.ListTableModel;
 import com.intellij.util.ui.UIUtil;
 import net.miginfocom.swing.MigLayout;
+import org.apache.commons.lang.StringUtils;
 import org.jetbrains.annotations.NotNull;
 import spp.jetbrains.PluginUI;
 import spp.jetbrains.UserData;
@@ -37,16 +39,23 @@ import spp.jetbrains.icons.PluginIcons;
 import spp.jetbrains.marker.service.ArtifactConditionService;
 import spp.jetbrains.marker.source.mark.api.SourceMark;
 import spp.jetbrains.marker.source.mark.inlay.InlayMark;
-import spp.jetbrains.plugin.LiveStatusManager;
+import spp.jetbrains.plugin.LiveStatusBarManager;
 import spp.jetbrains.marker.SourceMarkerKeys;
-import spp.jetbrains.sourcemarker.command.ui.status.config.LiveMeterConfigurationPanel;
+import spp.jetbrains.sourcemarker.command.status.ui.config.LiveBreakpointConfigurationPanel;
 import spp.jetbrains.sourcemarker.instrument.breakpoint.BreakpointHitColumnInfo;
-import spp.jetbrains.sourcemarker.status.util.AutocompleteField;
+import spp.jetbrains.sourcemarker.instrument.breakpoint.BreakpointHitWindowService;
+import spp.jetbrains.sourcemarker.command.util.AutocompleteField;
+import spp.jetbrains.sourcemarker.command.util.AutocompleteFieldRow;
 import spp.jetbrains.state.LiveStateBar;
+import spp.protocol.instrument.LiveBreakpoint;
 import spp.protocol.instrument.LiveInstrument;
 import spp.protocol.instrument.LiveSourceLocation;
-import spp.protocol.instrument.LiveSpan;
+import spp.protocol.instrument.event.LiveBreakpointHit;
 import spp.protocol.instrument.event.LiveInstrumentRemoved;
+import spp.protocol.instrument.event.TrackedLiveEvent;
+import spp.protocol.instrument.throttle.InstrumentThrottle;
+import spp.protocol.instrument.throttle.ThrottleStep;
+import spp.protocol.instrument.variable.LiveVariableControl;
 import spp.protocol.service.listen.LiveInstrumentListener;
 
 import javax.swing.*;
@@ -61,21 +70,27 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicReference;
+import java.util.function.Function;
+import java.util.stream.Collectors;
 
 import static spp.jetbrains.PluginBundle.message;
 import static spp.jetbrains.PluginUI.*;
 import static spp.jetbrains.utils.ViewUtils.addRecursiveMouseListener;
+import static spp.protocol.instrument.event.LiveInstrumentEventType.BREAKPOINT_REMOVED;
 
-public class SpanStatusBar extends JPanel implements LiveStateBar, VisibleAreaListener {
+public class BreakpointStatusBar extends JPanel implements LiveStateBar, LiveInstrumentListener, VisibleAreaListener {
 
     private final InlayMark inlayMark;
     private final LiveSourceLocation sourceLocation;
     private EditorImpl editor;
     private JWindow popup;
-    private LiveMeterConfigurationPanel configurationPanel;
+    private LiveBreakpointConfigurationPanel configurationPanel;
     private boolean disposed = false;
-    private final String placeHolderText = message("operation_name");
-    private LiveSpan liveSpan;
+    private final List<AutocompleteFieldRow> scopeVars;
+    private final Function<String, List<AutocompleteFieldRow>> lookup;
+    private final String placeHolderText = message("breakpoint_condition");
+    private LiveBreakpoint liveBreakpoint;
     private LiveBreakpointStatusPanel statusPanel;
     private JPanel wrapper;
     private JPanel panel;
@@ -83,22 +98,69 @@ public class SpanStatusBar extends JPanel implements LiveStateBar, VisibleAreaLi
     private boolean expanded = false;
     private final ListTableModel commandModel = new ListTableModel<>(
             new ColumnInfo[]{
-                    new BreakpointHitColumnInfo(message("meter_data")),
+                    new BreakpointHitColumnInfo(message("breakpoint_data")),
                     new BreakpointHitColumnInfo(message("time"))
             },
             new ArrayList<>(), 0, SortOrder.DESCENDING);
 
-    public SpanStatusBar(LiveSourceLocation sourceLocation, InlayMark inlayMark) {
+    public BreakpointStatusBar(LiveSourceLocation sourceLocation, List<String> scopeVars, InlayMark inlayMark) {
         this.sourceLocation = sourceLocation;
+        this.scopeVars = scopeVars.stream().map(it -> new AutocompleteFieldRow() {
+            public String getText() {
+                return it;
+            }
+
+            public String getDescription() {
+                return null;
+            }
+
+            public Icon getSelectedIcon() {
+                return PluginIcons.Nodes.variable;
+            }
+
+            public Icon getUnselectedIcon() {
+                return PluginIcons.Nodes.variable;
+            }
+        }).collect(Collectors.toList());
+        lookup = text -> scopeVars.stream()
+                .filter(v -> {
+                    String var = substringAfterLast(" ", text.toLowerCase());
+                    return !var.isEmpty() && !var.equalsIgnoreCase(v) && v.toLowerCase().contains(var);
+                }).map(it -> new AutocompleteFieldRow() {
+                    public String getText() {
+                        return it;
+                    }
+
+                    public String getDescription() {
+                        return null;
+                    }
+
+                    public Icon getSelectedIcon() {
+                        return PluginIcons.Nodes.variable;
+                    }
+
+                    public Icon getUnselectedIcon() {
+                        return PluginIcons.Nodes.variable;
+                    }
+                })
+                .limit(7)
+                .collect(Collectors.toList());
+
         this.inlayMark = inlayMark;
 
         initComponents();
         setupComponents();
     }
 
+    @Override
+    public void onInstrumentRemovedEvent(@NotNull LiveInstrumentRemoved event) {
+        LiveStatusBarManager.getInstance(inlayMark.getProject()).removeActiveLiveInstrument(event.getLiveInstrument());
+        this.liveBreakpoint = null;
+    }
+
     public void setLiveInstrument(LiveInstrument liveInstrument) {
-        this.liveSpan = (LiveSpan) liveInstrument;
-        setupAsActive();
+        this.liveBreakpoint = (LiveBreakpoint) liveInstrument;
+        setupAsActive(liveBreakpoint);
     }
 
     public void setWrapperPanel(JPanel wrapperPanel) {
@@ -107,7 +169,7 @@ public class SpanStatusBar extends JPanel implements LiveStateBar, VisibleAreaLi
 
     @Override
     public void visibleAreaChanged(VisibleAreaEvent e) {
-        spanOperationNameField.hideAutocompletePopup();
+        breakpointConditionField.hideAutocompletePopup();
         if (popup != null) {
             popup.dispose();
             popup = null;
@@ -119,8 +181,8 @@ public class SpanStatusBar extends JPanel implements LiveStateBar, VisibleAreaLi
     }
 
     public void focus() {
-        spanOperationNameField.grabFocus();
-        spanOperationNameField.requestFocusInWindow();
+        breakpointConditionField.grabFocus();
+        breakpointConditionField.requestFocusInWindow();
     }
 
     private void removeActiveDecorations() {
@@ -129,45 +191,54 @@ public class SpanStatusBar extends JPanel implements LiveStateBar, VisibleAreaLi
             closeLabel.setIcon(PluginIcons.close);
             configPanel.setBackground(getInputBackgroundColor());
 
-            if (!spanOperationNameField.getEditMode()) {
-                spanOperationNameField.setBorder(new CompoundBorder(
+            if (!breakpointConditionField.getEditMode()) {
+                breakpointConditionField.setBorder(new CompoundBorder(
                         new LineBorder(JBColor.DARK_GRAY, 0, true),
                         JBUI.Borders.empty(2, 6, 0, 0)));
-                spanOperationNameField.setBackground(PluginUI.getEditCompleteColor());
-                spanOperationNameField.setEditable(false);
+                breakpointConditionField.setBackground(PluginUI.getEditCompleteColor());
+                breakpointConditionField.setEditable(false);
             }
         });
     }
 
-    private void setupAsActive() {
-        LiveStatusManager.getInstance(inlayMark.getProject()).addStatusBar(inlayMark, new LiveInstrumentListener() {
+    private void setupAsActive(LiveBreakpoint liveBreakpoint) {
+        LiveStatusBarManager.getInstance(inlayMark.getProject()).addStatusBar(inlayMark, new LiveInstrumentListener() {
+            @Override
+            public void onBreakpointHitEvent(@NotNull LiveBreakpointHit event) {
+                if (statusPanel == null) return;
+                commandModel.insertRow(0, event);
+                statusPanel.incrementHits();
+            }
+
             @Override
             public void onInstrumentRemovedEvent(@NotNull LiveInstrumentRemoved event) {
                 if (statusPanel == null) return;
                 configLabel.setIcon(PluginIcons.eyeSlash);
 
                 if (event.getCause() == null) {
-                    statusPanel.setStatus("Complete", COMPLETE_COLOR_PURPLE);
+                    statusPanel.setStatus(message("complete"), COMPLETE_COLOR_PURPLE);
                 } else {
                     commandModel.insertRow(0, event);
-                    statusPanel.setStatus("Error", SELECT_COLOR_RED);
+                    statusPanel.setStatus(message("error"), SELECT_COLOR_RED);
                 }
             }
         });
         statusPanel = new LiveBreakpointStatusPanel();
-        statusPanel.setHitLimit(liveSpan.getHitLimit());
+        statusPanel.setHitLimit(liveBreakpoint.getHitLimit());
 
-        spanOperationNameField.setEditMode(false);
+        breakpointConditionField.setEditMode(false);
         removeActiveDecorations();
         configDropdownLabel.setVisible(false);
         SwingUtilities.invokeLater(() -> {
             mainPanel.removeAll();
             mainPanel.setLayout(new BorderLayout());
-            //statusPanel.setExpires(liveMeter.getExpiresAt());
+            statusPanel.setExpires(liveBreakpoint.getExpiresAt());
             mainPanel.add(statusPanel);
 
             remove(closeLabel);
-
+//                    JLabel searchLabel = new JLabel();
+//                    searchLabel.setIcon(PluginIcons.search);
+//                    add(searchLabel, "cell 2 0");
             expandLabel = new JLabel();
             expandLabel.setCursor(Cursor.getDefaultCursor());
             expandLabel.addMouseMotionListener(new MouseAdapter() {
@@ -191,6 +262,28 @@ public class SpanStatusBar extends JPanel implements LiveStateBar, VisibleAreaLi
                         table.setModel(commandModel);
                         table.setStriped(true);
                         table.setShowColumns(true);
+
+                        AtomicReference<LiveBreakpointHit> shownBreakpointHit = new AtomicReference<>();
+                        table.getSelectionModel().addListSelectionListener(event -> ApplicationManager.getApplication().invokeLater(() -> {
+                            if (table.getSelectedRow() > -1) {
+                                TrackedLiveEvent selectedEvent = (TrackedLiveEvent) commandModel.getItem(
+                                        table.convertRowIndexToModel(table.getSelectedRow())
+                                );
+                                if (selectedEvent.getEventType() == BREAKPOINT_REMOVED) return;
+                                LiveBreakpointHit selectedValue = (LiveBreakpointHit) selectedEvent;
+                                if (selectedValue.equals(shownBreakpointHit.getAndSet(selectedValue))) return;
+
+                                SwingUtilities.invokeLater(() -> {
+                                    expanded = false;
+                                    wrapper.remove(panel);
+
+                                    BreakpointHitWindowService.Companion.getInstance(inlayMark.getProject())
+                                            .clearContent();
+                                    BreakpointHitWindowService.Companion.getInstance(inlayMark.getProject())
+                                            .showBreakpointHit(shownBreakpointHit.get(), true);
+                                });
+                            }
+                        }));
 
                         table.setBackground(getBackgroundColor());
                         panel.add(scrollPane);
@@ -225,13 +318,26 @@ public class SpanStatusBar extends JPanel implements LiveStateBar, VisibleAreaLi
     }
 
     private void setupComponents() {
-        spanOperationNameField.setCanShowSaveButton(false);
-        spanOperationNameField.getDocument().addDocumentListener(new DocumentAdapter() {
+        JFormattedTextField spinnerTextField = ((JSpinner.NumberEditor) hitLimitSpinner.getEditor()).getTextField();
+        spinnerTextField.addKeyListener(new KeyAdapter() {
+            @Override
+            public void keyTyped(KeyEvent e) {
+                if (e.getKeyChar() == KeyEvent.VK_ENTER) {
+                    if (StringUtils.isNumeric(spinnerTextField.getText())) {
+                        ApplicationManager.getApplication().runWriteAction(() -> saveLiveBreakpoint());
+                    } else {
+                        spinnerTextField.setText("1");
+                    }
+                }
+            }
+        });
+        spinnerTextField.putClientProperty(UIUtil.HIDE_EDITOR_FROM_DATA_CONTEXT_PROPERTY, true);
+        breakpointConditionField.getDocument().addDocumentListener(new DocumentAdapter() {
             @Override
             protected void textChanged(DocumentEvent e) {
             }
         });
-        spanOperationNameField.addKeyListener(new KeyAdapter() {
+        breakpointConditionField.addKeyListener(new KeyAdapter() {
             @Override
             public void keyPressed(KeyEvent e) {
                 if (e.getKeyChar() == KeyEvent.VK_TAB) {
@@ -244,12 +350,12 @@ public class SpanStatusBar extends JPanel implements LiveStateBar, VisibleAreaLi
             public void keyTyped(KeyEvent e) {
                 if (e.getKeyChar() == KeyEvent.VK_ESCAPE) {
                     dispose();
-                } else if (e.getKeyChar() == KeyEvent.VK_ENTER && spanOperationNameField.getText().length() > 0) {
-                    ApplicationManager.getApplication().runWriteAction(() -> saveLiveSpan());
+                } else if (e.getKeyChar() == KeyEvent.VK_ENTER) {
+                    spinnerTextField.requestFocus();
                 }
             }
         });
-        spanOperationNameField.putClientProperty(UIUtil.HIDE_EDITOR_FROM_DATA_CONTEXT_PROPERTY, true);
+        breakpointConditionField.putClientProperty(UIUtil.HIDE_EDITOR_FROM_DATA_CONTEXT_PROPERTY, true);
 
         addMouseMotionListener(new MouseAdapter() {
             @Override
@@ -314,19 +420,19 @@ public class SpanStatusBar extends JPanel implements LiveStateBar, VisibleAreaLi
     }
 
     private void showConfigurationPopup(AtomicLong popupLastOpened) {
-        popup = new JWindow(SwingUtilities.getWindowAncestor(SpanStatusBar.this));
+        popup = new JWindow(SwingUtilities.getWindowAncestor(BreakpointStatusBar.this));
         popup.setType(Window.Type.POPUP);
         popup.setAlwaysOnTop(true);
 
         if (configurationPanel == null) {
-            configurationPanel = new LiveMeterConfigurationPanel(spanOperationNameField, inlayMark);
+            configurationPanel = new LiveBreakpointConfigurationPanel(breakpointConditionField);
         }
 
         popup.add(configurationPanel);
-        popup.setPreferredSize(new Dimension(SpanStatusBar.this.getWidth(), popup.getPreferredSize().height));
+        popup.setPreferredSize(new Dimension(BreakpointStatusBar.this.getWidth(), popup.getPreferredSize().height));
         popup.pack();
         popup.setLocation(configPanel.getLocationOnScreen().x - 1,
-                configPanel.getLocationOnScreen().y + SpanStatusBar.this.getHeight() - 2);
+                configPanel.getLocationOnScreen().y + BreakpointStatusBar.this.getHeight() - 2);
 
         popup.setVisible(true);
 
@@ -343,20 +449,36 @@ public class SpanStatusBar extends JPanel implements LiveStateBar, VisibleAreaLi
         });
     }
 
-    private void saveLiveSpan() {
-        spanOperationNameField.setShowSaveButton(false);
+    private void saveLiveBreakpoint() {
+        breakpointConditionField.setShowSaveButton(false);
 
         String condition = null;
-        Long expirationDate = null;
-        int hitLimit = -1;
+        if (!breakpointConditionField.getText().isEmpty()) {
+            condition = ArtifactConditionService.INSTANCE.getCondition(
+                    breakpointConditionField.getText(), inlayMark.getPsiElement()
+            );
+        }
+
+        LiveVariableControl variableControl = null;
+        long expirationDate = Instant.now().toEpochMilli() + (1000L * 60L * 15);
+        InstrumentThrottle throttle = InstrumentThrottle.Companion.getDEFAULT();
+        JFormattedTextField spinnerTextField = ((JSpinner.NumberEditor) hitLimitSpinner.getEditor()).getTextField();
+        int hitLimit = Integer.parseInt(spinnerTextField.getText());
         if (configurationPanel != null) {
-            if (configurationPanel.getCondition() != null) {
-                condition = ArtifactConditionService.INSTANCE.getCondition(
-                        configurationPanel.getCondition().getExpression(), inlayMark.getPsiElement()
+            expirationDate = Instant.now().toEpochMilli() + (1000L * 60L * configurationPanel.getExpirationInMinutes());
+            throttle = new InstrumentThrottle(
+                    configurationPanel.getRateLimitCount(),
+                    ThrottleStep.valueOf(configurationPanel.getRateLimitStep().toUpperCase())
+            );
+
+            if (configurationPanel.isVariableControlChanged()) {
+                variableControl = new LiveVariableControl(
+                        configurationPanel.getMaxObjectDepth(),
+                        configurationPanel.getMaxObjectSize(),
+                        configurationPanel.getMaxCollectionLength(),
+                        Collections.emptyMap(),
+                        Collections.emptyMap()
                 );
-            }
-            if (configurationPanel.getExpirationInMinutes() != -1) {
-                expirationDate = Instant.now().toEpochMilli() + (1000L * 60L * configurationPanel.getExpirationInMinutes());
             }
 
             configurationPanel.setNewDefaults();
@@ -365,8 +487,8 @@ public class SpanStatusBar extends JPanel implements LiveStateBar, VisibleAreaLi
         HashMap<String, String> meta = new HashMap<>();
         meta.put("original_source_mark", inlayMark.getId());
 
-        LiveSpan instrument = new LiveSpan(
-                spanOperationNameField.getText(),
+        LiveBreakpoint instrument = new LiveBreakpoint(
+                variableControl,
                 sourceLocation,
                 condition,
                 expirationDate,
@@ -375,17 +497,14 @@ public class SpanStatusBar extends JPanel implements LiveStateBar, VisibleAreaLi
                 false,
                 false,
                 false,
-                null,
+                throttle,
                 meta
         );
-        UserData.liveInstrumentService(inlayMark.getProject()).addLiveInstrument(instrument).onComplete(it -> {
+        UserData.liveInstrumentService(inlayMark.getProject()).addLiveBreakpoint(instrument).onComplete(it -> {
             if (it.succeeded()) {
-                liveSpan = (LiveSpan) it.result();
-                LiveStatusManager.getInstance(inlayMark.getProject()).addActiveLiveInstrument(liveSpan);
-
-                ApplicationManager.getApplication().invokeLater(() -> {
-                    inlayMark.dispose(); //dispose this bar
-                });
+                liveBreakpoint = it.result();
+                inlayMark.putUserData(SourceMarkerKeys.getINSTRUMENT_ID(), liveBreakpoint.getId());
+                LiveStatusBarManager.getInstance(inlayMark.getProject()).addActiveLiveInstrument(liveBreakpoint);
             } else {
                 it.cause().printStackTrace();
             }
@@ -404,15 +523,21 @@ public class SpanStatusBar extends JPanel implements LiveStateBar, VisibleAreaLi
         List<SourceMark> groupedMarks = inlayMark.getUserData(SourceMarkerKeys.getGROUPED_MARKS());
         if (groupedMarks != null) groupedMarks.forEach(SourceMark::dispose);
 
-        if (liveSpan != null) {
-            UserData.liveInstrumentService(inlayMark.getProject()).removeLiveInstrument(liveSpan.getId()).onComplete(it -> {
+        if (liveBreakpoint != null) {
+            UserData.liveInstrumentService(inlayMark.getProject()).removeLiveInstrument(liveBreakpoint.getId()).onComplete(it -> {
                 if (it.succeeded()) {
-                    LiveStatusManager.getInstance(inlayMark.getProject()).removeActiveLiveInstrument(liveSpan);
+                    LiveStatusBarManager.getInstance(inlayMark.getProject()).removeActiveLiveInstrument(liveBreakpoint);
                 } else {
                     it.cause().printStackTrace();
                 }
             });
         }
+    }
+
+    public static String substringAfterLast(String delimiter, String value) {
+        int index = value.lastIndexOf(delimiter);
+        if (index == -1) return value;
+        else return value.substring(index + delimiter.length());
     }
 
     private void initComponents() {
@@ -422,7 +547,9 @@ public class SpanStatusBar extends JPanel implements LiveStateBar, VisibleAreaLi
         configLabel = new JLabel();
         configDropdownLabel = new JLabel();
         mainPanel = new JPanel();
-        spanOperationNameField = new AutocompleteField(inlayMark.getProject(), placeHolderText, Collections.emptyList(), null, inlayMark.getArtifactQualifiedName(), false);
+        breakpointConditionField = new AutocompleteField(inlayMark.getProject(), placeHolderText, scopeVars, lookup, inlayMark.getArtifactQualifiedName(), false);
+        label1 = new JLabel();
+        hitLimitSpinner = new JBIntSpinner(1, 1, 10_000);
         timeLabel = new JLabel();
         separator1 = new JSeparator();
         closeLabel = new JLabel();
@@ -434,7 +561,7 @@ public class SpanStatusBar extends JPanel implements LiveStateBar, VisibleAreaLi
         setLayout(new MigLayout(
             "hidemode 3",
             // columns
-            "0[fill]" +
+            "0[fill]0" +
             "[grow,fill]" +
             "[fill]",
             // rows
@@ -455,7 +582,7 @@ public class SpanStatusBar extends JPanel implements LiveStateBar, VisibleAreaLi
                 "[grow]"));
 
             //---- configLabel ----
-            configLabel.setIcon(PluginIcons.spanConfig);
+            configLabel.setIcon(PluginIcons.breakpointConfig);
             configPanel.add(configLabel, "cell 0 0");
 
             //---- configDropdownLabel ----
@@ -475,14 +602,25 @@ public class SpanStatusBar extends JPanel implements LiveStateBar, VisibleAreaLi
                 // rows
                 "0[grow]0"));
 
-            //---- spanOperationNameField ----
-            spanOperationNameField.setBackground(getInputBackgroundColor());
-            spanOperationNameField.setBorder(new CompoundBorder(
+            //---- breakpointConditionField ----
+            breakpointConditionField.setBackground(getInputBackgroundColor());
+            breakpointConditionField.setBorder(new CompoundBorder(
                     new LineBorder(UIUtil.getBoundsColor(), 1, true),
                     JBUI.Borders.empty(2, 6, 0, 0)));
-            spanOperationNameField.setFont(BIG_FONT);
-            spanOperationNameField.setMinimumSize(new Dimension(0, 27));
-            mainPanel.add(spanOperationNameField, "cell 0 0 2 1");
+            breakpointConditionField.setFont(BIG_FONT);
+            breakpointConditionField.setMinimumSize(new Dimension(0, 27));
+            mainPanel.add(breakpointConditionField, "cell 0 0");
+
+            //---- label1 ----
+            label1.setText(message("hit_limit"));
+            label1.setForeground(Color.gray);
+            label1.setFont(SMALLER_FONT);
+            mainPanel.add(label1, "cell 1 0");
+
+            //---- hitLimitSpinner ----
+            hitLimitSpinner.setBackground(null);
+            //hitLimitSpinner.setModel(new SpinnerNumberModel(1, 1, null, 1));
+            mainPanel.add(hitLimitSpinner, "cell 1 0");
 
             //---- timeLabel ----
             timeLabel.setIcon(PluginIcons.clock);
@@ -512,7 +650,9 @@ public class SpanStatusBar extends JPanel implements LiveStateBar, VisibleAreaLi
     private JLabel configLabel;
     private JLabel configDropdownLabel;
     private JPanel mainPanel;
-    private AutocompleteField spanOperationNameField;
+    private AutocompleteField<AutocompleteFieldRow> breakpointConditionField;
+    private JLabel label1;
+    private JBIntSpinner hitLimitSpinner;
     private JLabel timeLabel;
     private JSeparator separator1;
     private JLabel closeLabel;
