+++ plugin/src/main/kotlin/spp/jetbrains/sourcemarker/command/status/ui/LiveBreakpointStatusPanel.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package spp.jetbrains.sourcemarker.command.ui.status;
+package spp.jetbrains.sourcemarker.command.status.ui;
 
 import com.codahale.metrics.Meter;
 import com.jgoodies.forms.factories.FormFactory;
