+++ plugin/src/main/kotlin/spp/jetbrains/sourcemarker/command/status/ui/LogStatusBar.jfd
