+++ plugin/src/main/kotlin/spp/jetbrains/sourcemarker/command/status/ui/MeterStatusBar.jfd
