+++ plugin/src/main/kotlin/spp/jetbrains/sourcemarker/command/status/ui/config/LiveLogConfigurationPanel.jfd
