+++ plugin/src/main/kotlin/spp/jetbrains/sourcemarker/command/status/ui/element/AutocompleteDropdown.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package spp.jetbrains.sourcemarker.command.ui.status.element;
+package spp.jetbrains.sourcemarker.command.status.ui.element;
 
 import com.intellij.openapi.project.Project;
 import io.vertx.core.Vertx;
